+++ src/main/java/roomescape/ReservationController.java
@@ -1,34 +1,33 @@
 package roomescape;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
 
-@Controller
+@RestController
 public class ReservationController {
+
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
 
     private List<Reservation> reservations = new ArrayList<>();
     private AtomicLong index = new AtomicLong(0);
 
-    @GetMapping("/")
-    public String showHomePage() {
-        return "home";
-    }
-
-    @GetMapping("/reservation")
-    public String showReservationPage() {
-        return "reservation";
-    }
-
     @GetMapping("/reservations")
-    @ResponseBody
     public List<Reservation> getReservations() {
-        return reservations;
+        return jdbcTemplate.query("SELECT id, name, date, time FROM reservation",
+                (rs, rowNum) -> new Reservation (
+                        rs.getLong("id"),
+                        rs.getString("name"),
+                        rs.getString("date"),
+                        rs.getString("time")
+                ));
     }
 
     @PostMapping("/reservations")
@@ -47,7 +46,6 @@ public class ReservationController {
                 .body(reservation);
     }
 
-
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> cancelReservation(@PathVariable long id) {
         boolean removed = reservations.removeIf(reservation -> reservation.getId() == id);
@@ -56,5 +54,4 @@ public class ReservationController {
         }
         return ResponseEntity.noContent().build();
     }
-
 }
