+++ src/main/java/roomescape/ViewController.java
@@ -1,0 +1,17 @@
+package roomescape;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+@Controller
+public class ViewController {
+    @GetMapping("/")
+    public String showHomePage() {
+        return "home";
+    }
+
+    @GetMapping("/reservation")
+    public String showReservationPage() {
+        return "reservation";
+    }
+
+}
